Move ConfigParserBase to new file
